requirements.txt
Django==4.0.6
djangorestframework==3.14.0
PyMySQL==1.1.0